Ref: extracting formatDuration() helper.
